Fix offsetting to start of buffer on empty offset function
If arg FNC is nil, the function would return pos 0 - BEG, which is the
start of the buffer.

If arg FNC is empty, we want some sensible default behavior. BEG is the
start of the beginning match. Usually we want it to remain visible, and
start folding right after it, so offset length(MATCH) sounds sensible.